+++ src/prime_rl/utils/validation.py
@@ -4,11 +4,11 @@ from typing import Optional
 
 from prime_rl.configs.inference import InferenceConfig
 from prime_rl.configs.orchestrator import OrchestratorConfig
-from prime_rl.configs.rl_trainer import RLTrainerConfig
+from prime_rl.configs.trainer import TrainerConfig
 
 
 def validate_shared_ckpt_config(
-    trainer: RLTrainerConfig,
+    trainer: TrainerConfig,
     orchestrator: OrchestratorConfig,
 ) -> None:
     if trainer.ckpt and not orchestrator.ckpt:
@@ -30,7 +30,7 @@ def validate_shared_ckpt_config(
 
 
 def validate_shared_model_name(
-    trainer: RLTrainerConfig,
+    trainer: TrainerConfig,
     orchestrator: OrchestratorConfig,
     inference: Optional[InferenceConfig] = None,
 ) -> None:
@@ -48,7 +48,7 @@ def validate_shared_model_name(
 
 
 def validate_shared_output_dir(
-    trainer: RLTrainerConfig,
+    trainer: TrainerConfig,
     orchestrator: OrchestratorConfig,
 ) -> None:
     if trainer.output_dir != orchestrator.output_dir.parent:
@@ -58,7 +58,7 @@ def validate_shared_output_dir(
 
 
 def validate_shared_wandb_config(
-    trainer: RLTrainerConfig,
+    trainer: TrainerConfig,
     orchestrator: OrchestratorConfig,
 ) -> None:
     if trainer.wandb and not orchestrator.wandb:
@@ -81,7 +81,7 @@ def validate_shared_wandb_config(
 
 
 def validate_shared_max_steps(
-    trainer: RLTrainerConfig,
+    trainer: TrainerConfig,
     orchestrator: OrchestratorConfig,
 ) -> None:
     if trainer.max_steps != orchestrator.max_steps:
@@ -91,7 +91,7 @@ def validate_shared_max_steps(
 
 
 def validate_shared_max_async_level(
-    trainer: RLTrainerConfig,
+    trainer: TrainerConfig,
     orchestrator: OrchestratorConfig,
 ) -> None:
     if trainer.max_async_level != orchestrator.max_async_level:
@@ -101,7 +101,7 @@ def validate_shared_max_async_level(
 
 
 def validate_shared_weight_broadcast(
-    trainer: RLTrainerConfig,
+    trainer: TrainerConfig,
     orchestrator: OrchestratorConfig,
     inference: Optional[InferenceConfig] = None,
 ) -> None:
