+++ tests/unit/train/rl/test_loss.py
@@ -1,7 +1,7 @@
 import pytest
 import torch
 
-from prime_rl.configs.rl_trainer import CustomLossConfig, LossConfig
+from prime_rl.configs.trainer import CustomLossConfig, LossConfig
 from prime_rl.trainer.rl.loss import LossInputs, LossOutputs, compute_entropy, compute_loss, setup_loss_fn
 
 pytestmark = [pytest.mark.gpu]
